Refactor code for displaying MoneyRequestReportView
